LayoutTests/fast/parser/fragment-parser-expected.txt

Test that fragment parsing does not affect the host document.

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".


PASS parse('<span><body bgcolor=red>') is ['<span></span>','<html><body></body></html>']
PASS parse('<span><html bgcolor=red>') is ['<span></span>','<html><body></body></html>']
FAIL parse('<span><meta>') should be <span></span>,<html><body></body></html>. Was <span><meta></span>,<html><body></body></html>.
FAIL parse('<span><base>') should be <span></span>,<html><body></body></html>. Was <span><base></span>,<html><body></body></html>.
FAIL parse('<html><script>') should be ,<html><body></body></html>. Was <script></script>,<html><body></body></html>.
FAIL parse('<html><style>') should be ,<html><body></body></html>. Was <style></style>,<html><body></body></html>.
FAIL parse('<html><meta>') should be ,<html><body></body></html>. Was <meta>,<html><body></body></html>.
FAIL parse('<html><link>') should be ,<html><body></body></html>. Was <link>,<html><body></body></html>.
PASS parse('<html><object>') is ['<object></object>','<html><body></body></html>']
PASS parse('<html><embed>') is ['<embed>','<html><body></body></html>']
FAIL parse('<html><title>') should be ,<html><body></body></html>. Was <title></title>,<html><body></body></html>.
FAIL parse('<html><isindex>') should be <div><hr>This is a searchable index. Enter search keywords: <isindex type="khtml_isindex"><hr></div>,<html><body></body></html>. Was <form><hr><label>This is a searchable index. Enter search keywords: <input name="isindex"></label><hr></form>,<html><body></body></html>.
FAIL parse('<html><base>') should be ,<html><body></body></html>. Was <base>,<html><body></body></html>.
PASS parse('<html><div>') is ['<div></div>','<html><body></body></html>']
FAIL parse('<frameset>') should be <frameset></frameset>,<html><body></body></html>. Was ,<html><body></body></html>.
PASS parse('<html>x', true) is ['x','no document element']
PASS successfullyParsed is true

TEST COMPLETE
